fix: error directory rename
