+++ src/utils/index.ts
@@ -2,7 +2,7 @@ export * from './artifact-v2';
 export * from './available-rpc';
 export * from './compare';
 export * from './fallback-provider';
-export * from './errors';
+export * from './error';
 export * from './format';
 export * from './gas';
 export * from './network';
